WORKFLOWEDITRECORD-7 : The history is not removed when a record is removed
git-svn-id: http://dev.lutece.paris.fr/svn/lutece/portal/trunk/plugins/directory-form-workflow/module-workflow-editrecord@35594 bab10101-e421-0410-a517-8ce0973de3ef

+++ src/java/fr/paris/lutece/plugins/workflow/modules/editrecord/service/EditRecordService.java
@@ -221,8 +221,12 @@ public final class EditRecordService
     public void removeByIdHistory( int nIdHistory, int nIdTask )
     {
         EditRecord editRecord = find( nIdHistory, nIdTask );
-        _editRecordValueService.remove( editRecord.getIdHistory(  ) );
-        EditRecordHome.removeByIdHistory( nIdHistory, nIdTask );
+
+        if ( editRecord != null )
+        {
+            _editRecordValueService.remove( editRecord.getIdHistory(  ) );
+            EditRecordHome.removeByIdHistory( nIdHistory, nIdTask );
+        }
     }
 
     /**
